@@ -1,5 +1,10 @@
+<script>
+	/** @type {{children?: import('svelte').Snippet}} */
+	let { children } = $props();
+</script>
+
 <svelte:head>
 	<title>Email and password example with 2FA and WebAuthn in SvelteKit</title>
 </svelte:head>
 
-<slot />
+{@render children?.()}
